amdgpu, ps top metrics
